refactor(extraction): split extraction internals
